pyproject.toml
[project]
name = "predection-model"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.26.0",
    "pyyaml>=6.0.3",
    "python-binance>=1.0.19",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "numpy>=1.26",
    "lightgbm>=4.3",
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning:sklearn.utils.validation",
    "ignore::DeprecationWarning:websockets",
    "ignore::DeprecationWarning:binance",
]
